% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{add_footnote}
\alias{add_footnote}
\title{Add footnote to a flextable}
\usage{
add_footnote(
  x,
  str,
  symb = "",
  i = NULL,
  j = NULL,
  size = 8,
  part = "body",
  inline = TRUE,
  sep = "\\n"
)
}
\arguments{
\item{x}{A flextable}

\item{str}{The footnote text (a character string).}

\item{symb}{Footnote symbol (Default: "")}

\item{i}{Row}

\item{j}{Column}

\item{size}{Text size (Default: 8)}

\item{part}{Part of flextable (Default: "body")}

\item{inline}{whether to add footnote on same line as previous footnote or not (Default: TRUE)}

\item{sep}{character string to use as a separator between footnotes (Default: "\\n")}
}
\value{
A flextable
}
\description{
Add footnote to a flextable
}
