% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{add_headers}
\alias{add_headers}
\title{Add header rows}
\usage{
add_headers(x, headers = NULL, colspan = NULL)
}
\arguments{
\item{x}{A flextable}

\item{headers}{A list of character vectors (or a single character vector).}

\item{colspan}{A list of numeric vectors (or a single numeric vector).}
}
\description{
Add header rows
}
