% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{colwidths}
\alias{colwidths}
\title{Control all column widths in a flextable}
\usage{
colwidths(x, widths, unit = "mm")
}
\arguments{
\item{x}{A flextable}

\item{widths}{A numeric vector of widths}

\item{unit}{Unit for widths, one of "in", "cm", "mm" (Default: "mm")}
}
\value{
A flextable.
}
\description{
Control all column widths in a flextable
}
