% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{marker}
\alias{marker}
\title{Convenience function for common footnote markers}
\usage{
marker(x = "asterisk")
}
\arguments{
\item{x}{"asterisk", "dagger", "ddagger", or "silcrow" (Default: "asterisk)}
}
\value{
A character string
}
\description{
Convenience function for common footnote markers
}
