% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{style_table}
\alias{style_table}
\title{Apply overall style to a flextable}
\usage{
style_table(x)
}
\arguments{
\item{x}{A flextable}
}
\value{
A flextable
}
\description{
Apply overall style to a flextable
}
