% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{underline_header}
\alias{underline_header}
\title{Underline column headings in a flextable}
\usage{
underline_header(x, i = 1, j)
}
\arguments{
\item{x}{A flextable}

\item{i}{Row}

\item{j}{Column(s)}
}
\value{
A flextable
}
\description{
Underline column headings in a flextable
}
