% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{add_headers}
\alias{add_headers}
\title{Add header rows}
\usage{
add_headers(x, v = NULL, c = NULL)
}
\arguments{
\item{x}{A flextable}

\item{headers}{A list of}
}
\value{

}
\description{
Add header rows
}
