% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnish-package.R
\docType{package}
\name{varnish-package}
\alias{varnish}
\alias{varnish-package}
\title{The varnish package}
\description{
This package does not contain any useful R code. It serves to hold CSS, JS,
and HTML styling for Carpentries lessons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carpentries.github.io/varnish/}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@carpentries.org} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

Other contributors:
\itemize{
  \item François Michonneau \email{francois@carpentries.org} [contributor]
}

}
\keyword{internal}
