% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{varnish}
\alias{varnish}
\title{Varnish a table}
\usage{
varnish(
  x,
  v = NULL,
  c = NULL,
  underline_i = 1,
  underline_j = NULL,
  widths = NULL,
  unit = "mm"
)
}
\arguments{
\item{unit}{}
}
\value{

}
\description{
Varnish a table
}
